backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-dotenv>=1.0
# Production extras
gunicorn>=21.2
whitenoise>=6.7
dj-database-url>=2.2
psycopg2-binary>=2.9
